Add files via upload
This spreadsheet works just the same way as the example
shown in the Bombers and Seagulls
Spreadsheet (and the video) but has 10,000 rows of data instead of 20.  It is designed to provide a realistic
simulation of the cost-benefit assumptions that must be made to set the classification
threshold for a medical diagnostic product. You can download the spreadsheet by clicking on, "Cancer Diagnosis.xlsx."
The ranked scores
in Column A are the level of a certain protein as measured by the diagnostic
test. 

The true Condition for each protein level score is given in
Column C [Cancer =1, No Cancer = 0].

A threshold for positive classification can be set between
any two protein levels. For each threshold given in Column A, the resulting number
of False Negative classification errors is given in the same row of Column H,
and the False Positive classification errors in the same row of Column F.

This spreadsheet is designed to allow you to observe how
changing cost inputs impacts both (a) the overall costs of using a cancer diagnostic
test at each threshold, and (2) what threshold should be chosen as optimal
(minimizing cost).

The cells to input the assumed costs of classification
errors are Cell G3 for cost per False Negative (missing a case of cancer) and
H3 for cost per False Positive (a false alarm). 
Total costs at each threshold are given in Column K.

The minimum total cost, and minimum cost per event (per
diagnostic classification reported), are displayed in cells K4 and L4, and the optimum
threshold – the lowest protein level score that should be classified positive -
is displayed in cell M4.

At the default costs of $50,000 per False Negative error and
$500 per False Positive error, the minimum cost per event is $119.90, and the optimal
threshold for positive classification is 16551.930.

Try changing the inputs yourself. If you keep the cost per
False Negative the same, but raise the cost per False Positive, would you in
general expect the new optimum threshold to be higher (fewer total positive
classifications) or lower (more total positive classifications)? 

Change the cost per False Positive to $1500 and you will see
that the optimum threshold is higher.
The threshold for the first Positive classification moves from 16551.930 [item
ranked 2094] to 16824.137 [item ranked 1822].

Similarly, if you reset the cost per FP at $500 and raise
the cost per FN to from $50,000 to $500,000, the optimal threshold is lower –
it falls to 13307.537 [item ranked 5338]