Add files via upload
Binary Performance
This Spreadsheet is a calculator that starts with the
Confusion matrix as an input,  and
outputs  both the eight most important binary
classification performance metrics and the one specific (x,y) coordinate – a
point on an ROC Curve  - that summarizes
the Confusion Matrix’s performance.

Binary Performance Metrics

Input
any value between 0 and 1 for the probability of True Positive classifications
– the joint probability labeled letter e in the confusion matrix – into cell
G9.

Next,
enter any valid values (that is, any values less than or equal to the value
entered into cell G9) for the condition incidence – letter a - (cell E9) and
the “classification incidence” (also called “test incidence”) – letter c -
(cell G7).

The
Excel Spreadsheet will output:

 -Values
for the remaining five cells of the Confusion Matrix, designated by letters b,
d, f, g, and h.

-Values
for the eight primary performance metrics used to evaluate binary classification
systems, Column F, rows 36-39 and 41-44. 

 -One
point on the ROC Curve corresponding to this particular Confusion Matrix.  Recall that the x-axis coordinate equals the
False Positive Rate – output in Cell N37 – and the y axis coordinate equals the
true positive rate – output in Cell O37.

Assume
the Confusion Matrix shows the performance of a model that tries to predict which
visitors to an Automobile dealership will buy a car from that dealer.

Experiment
with how a change in one or more inputs impacts the outputs. For example,
change the probability that an event will be a True Positive [Cell G9] to .18,
without changing the condition incidence or test incidence. 

Question:
What is the conditional probability that a visitor classified as Positive by
the predictive model (the “test”) will buy a car (the condition)?    

Answer:  90%.  The
answer is the “Positive Predictive Value” – also written as the conditional
probability p ("+" | Test POS) – as is found in
cell F41, is 0.9. 

Question: What point on the ROC Curve summarizes this confusion
matrix performance? Answer:

Answer: (.03, .6). The value 3% is the False Positive rate – also written
as the conditional probability p(Test POS | "-") – found in Cells N37
and F38.  The value 60% is the true
positive rate – also known as the conditional probability p(Test POS |
"+")  - Given in cells O37 and
F36.

Question: What is the probability that a visitor classified as
“Negative” by the predictive model will not buy 
a car from that dealer?


Answer: 85%. This value is the Negative Predictive Value, also written as the
conditional probability p("-" | Test NEG) – found in Cell F44.