Add files via upload
This Spreadsheet can be used to practice sorting data so that it
is ordered by the “score” used for predictive purposes.  Copy cells B18 to G30 and “paste special:
values and number formats.” Past into the upper left hand corner of a new
spreadsheet. Then select the right-most Column – containing the sum of
standardized heights and weights – and choose Data/Sort/Descending – when
asked, choose “Expand the selection.” You will now have created a ranked set of
scores and outcomes, just like what is already provided in the Bombers and Seagulls and Cancer Diagnostics Spreadsheets.  Repeat the process but sort on height alone,
on weight alone, or on a combination of height and weight not previously
standardized (Column G, rows 2 – 13) to compare the performance of the various
scores when used for binary classification.